move mass cmd to slash cmd
